moved stuff, adressed comments
